requirements.txt
pillow==8.0.0
datasets==2.14.6
tqdm==4.66.5